% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coxmulti.R
\name{coxmulti}
\alias{coxmulti}
\title{Multivariate cox proportional hazard data analysis}
\usage{
coxmulti(
  C1 = NULL,
  C2 = NULL,
  C3 = NULL,
  C4 = NULL,
  C5 = NULL,
  survdur,
  event,
  data
)
}
\arguments{
\item{C1}{Covar1}

\item{C2}{Covar2}

\item{C3}{Covar3}

\item{C4}{Covar4}

\item{C5}{Covar5}

\item{survdur}{"Column/Variable name" consisting duration of survival.}

\item{event}{"Column/Variable name" consisting survival event.}

\item{data}{High dimensional data containing survival observations and covariates.}
}
\value{
Data set containing the survival estimates and Pvalue.
}
\description{
Given the dimensions of the variables and survival informations. The function
performs multivariate Cox PH by taking 5 variables at a time.
}
\examples{
##
coxmulti(C1="PGC",C2="C7",C3="HPN",C4="DDC",C5=NULL,survdur="os",event="death",data=hnscc2)
}
