% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/srdata.R
\docType{data}
\name{srdata}
\alias{srdata}
\title{protein data}
\format{
A dataframe with 288 rows and 252 columns
\describe{
\item{ID}{ID of subjects}
\item{Visit}{Number of times observations recorded}
\item{event}{death as event 1 if died or 0 if alive}
\item{OS}{Duration of overall survival}
\item{leftcensored}{Left censoring information}
\item{lc}{Left censoring information}
\item{C6kine,.....,GFRalpha4}{High dimensional covariates}}
}
\usage{
data(srdata)
}
\description{
High dimensional pprotein gene expression data
}
\examples{
data(srdata)
}
\keyword{datasets}
