% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unihdcoxma.R
\name{unihdcoxma}
\alias{unihdcoxma}
\title{High dimensional univariate cox proportional hazard model.}
\usage{
unihdcoxma(m, n, survdur, event, sig, ths, b, d, data)
}
\arguments{
\item{m}{Starting column number from where high dimensional variates to be selected.}

\item{n}{Ending column number till where high dimensional variates to be selected.}

\item{survdur}{"Column/Variable name" consisting duration of survival.}

\item{event}{"Column/Variable name" consisting survival event.}

\item{sig}{Level of significance pre-determined by the user.}

\item{ths}{A numeric between 0 to 100.}

\item{b}{Number of MCMC iterations to burn.}

\item{d}{Number of draws for the iterations.}

\item{data}{High dimensional data containing survival observations and high dimensional covariates.}
}
\value{
Data frame containing the beta and alpha values of active variables among the significant variables.
}
\description{
Given the dimension of variables and survival information risks the function
filters significant variables, allowing the user to fit univariate COx PH model. Further, it performs mediation
analysis among the significant variables and provides handful variables with their alpha.a values
which are mediator model exposure coefficients and beta.a coefficients.
}
\examples{
##
data(hnscc)
unihdcoxma(m=8,n=105,survdur="os",event="death",sig=0.5,ths=0.02,b=1000,d=10,data=hnscc2)
##
}
